Homework 1-2 Submission Candidate
Is that really IT?! "Happy Thanksgiving" after the horror that was HW 1-1? But HW 1-1 is the one that's due at the end of Thanksgiving weekend, not 1-2!

And to top it all off, the auto-grader is bugged.  Well, what can you expect from a free online course? 6.002x just set the bar too high, that's all...